requirements.txt
# paho-mqtt<2.0.0
imageio<3.0.0
numpy==1.12.1 # Avoid upgrading, this version is already on my Raspberry Pis
python-nmap<1.0.0
spur<1.0.0
